+++ packages/android_alarm_manager_plus/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 2.0.5
+
+- Fix embedding issue in example
+- Update dependencies in example
+
+## 2.0.4
+
+- Fix FlutterEngine initialisation to avoid crashes when app is closed and plugin tries to schedule an alarm.
+
 ## 2.0.3
 
 - Handle Android 12 behavior changes for exact alarms scheduling