+++ packages/battery_plus/battery_plus_linux/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.1.1
+
+- Update dbus to 0.7.1
+
+## 1.1.0
+
+- Add batteryState getter
+
 ## 1.0.4
 
 - Update Flutter dependencies