+++ packages/connectivity_plus/connectivity_plus_linux/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.3.0
+
+- Bump nm plugin to 0.5.0
+
+## 1.2.0
+
+- Add bluetooth as connectivity result
+
 ## 1.1.1
 
 - Dependencies update